# openapi.yaml
openapi: 3.1.0
info:
  title: Programmable Monitoring Platform API
  description: These sets of APIs describes the principal functionalities of the ROBUST-6G PMP.
  version: 1.0.0
...
servers:
  - url: http://localhost:5000
    description: Local / default environmen
paths:
  /ConfigurationManager/DeployNetworkTool:
    post:
      tags:
        - ConfigurationManager
      summary: "Security Orchestrator - Request to deploy a network tool"
      description: "This endpoint allows the NXW Security Orchestrator to request the deployment of a network tool"
      operationId: "deployNetworkTool"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                podIP:
                  type: string
                  description: "IP address of the pod where the tool will be deployed"
                interface:
                  type: string
                  description: "Network interface on which the tool will sniff traffic"
                toolName:
                  type: string
                  description: "Name of the network tool to deploy"
                sigma_rules:
                  type: array
                  items:
                    type: string
                    description: "Sigma rules generated from the SSLA by the Security Orchestrator."
                configuration:
                  type: array
                  items:
                    type: string
                    description: "Configuration in JSON format from the ZTSO"
                  examples:
                    - 
                      - TSHARK_BASE_TOPIC: tshark_traces
              required:
                - podIP
                - interface
                - toolName
      responses:
        '200':
          description: "Request successfully. You can collect the information in IP:Port -> kafka_robust6g-node1.lan:9094"
  /ConfigurationManager/DeployInfrastructureTool: 
    post:
      tags:
        - ConfigurationManager
      summary: "Security Orchestrator - Request to deploy a infrastructure tool"
      description: "This endpoint allows the NXW Security Orchestrator to request the deployment of a infrastructure tool"
      operationId: "deployInfrastructureTool"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                podIP:
                  type: string
                  description: "IP address of the pod where the tool will be deployed"
                toolName:
                  type: string
                  description: "Name of the infrastructure tool to deploy"
                sigma_rules:
                  type: array
                  items:
                    type: string
                    description: "Sigma rules generated from the SSLA by the Security Orchestrator."
              required:
                - podIP
                - toolName
                - sigma_rules
      responses:
        '200':
          description: OK
  /ConfigurationManager/DeployServiceTool: 
    post:
      tags:
        - ConfigurationManager
      summary: "Security Orchestrator - Request to deploy a service tool"
      description: "This endpoint allows the NXW Security Orchestrator to request the deployment of a service tool"
      operationId: "deployServiceTool"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                podIP:
                  type: string
                  description: "IP address of the pod where the tool will be deployed"
                toolName:
                  type: string
                  description: "Name of the service tool to deploy"
                sigma_rules:
                  type: array
                  items:
                    type: string
                    description: "Sigma rules generated from the SSLA by the Security Orchestrator."
              required:
                - podIP
                - toolName
                - sigma_rules
      responses:
        '200':
          description: OK
  /ConfigurationManager/DeploySecurityTool: 
    post:
      tags:
        - ConfigurationManager
      summary: "Security Orchestrator - Request to deploy a security tool"
      description: "Deploy snort3 with optional configuration overrides and optional custom rules."
      operationId: "deploySecurityTool"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                configuration:
                  type: object
                  additionalProperties: true
                  description: "Optional snort3 env var overrides."
                  example:
                    SNORT_KAFKA_TOPIC_OUT: "snort_alerts"
                rules:
                  type: array
                  items:
                    type: string
                  description: "Optional list of custom Snort3 rules. Each rule must include a sid."
                  example:
                    - "alert tcp any any -> any any (msg:\"snort3 test rule\"; sid:10001; rev:1;)"
                include_default_rules:
                  type: boolean
                  description: "If true, snort3 uses community rules together with the provided custom rules. If omitted while rules are sent, it is interpreted as false."
      responses:
        '200':
          description: "Request successfully. You can collect the information in IP:Port -> kafka_robust6g-node1.lan:9094"
  /ConfigurationManager/getConfigurationOptions:
    get:
      tags:
        - ConfigurationManager
      summary: "Security Orchestrator - Requests to know the different configurations that a tool can manage"
      description: "This endpoint allows the NXW Security Orchestrator to know what options it has for configuring a tool"
      operationId: "getConfigurationOptions"
      parameters:
        - name: toolName
          in: query
          description: "Name of the tool for which the orchestrator wants to know the options"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Options of the tool"
          content:
            application/json:
              schema:
                type: object
                properties:
                  configurationOptions:
                    type: array
                    description: "List of options. Can be one or more options"
                    items:
                      type: string
                    examples:
                      - "memory"
                      - "disk"
                      - "cpu"
                      - "sensors"
                      - "internet speed"
  /ConfigurationManager/getConfiguration:
    get:
      tags:
        - ConfigurationManager
      summary: "Security Orchestrator - Retrieve a persisted deployment configuration"
      description: "Return the current stored configuration for a deployment, including revision metadata and, for snort3, the current rules_config."
      operationId: "getConfiguration"
      parameters:
        - name: config_id
          in: query
          description: "The stable configuration identifier returned by a previous deploy call."
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Current persisted configuration"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  config_id:
                    type: string
                  data:
                    type: object
                    properties:
                      endpoint:
                        type: string
                      tool_name:
                        type: string
                      created_at:
                        type: string
                      updated_at:
                        type: string
                      revision:
                        type: integer
                      current_version_hash:
                        type: string
                      resolved_env:
                        type: object
                        additionalProperties:
                          type: string
                      rules_config:
                        type: object
                        description: "Present for snort3 only."
                        properties:
                          include_default_rules:
                            type: boolean
                          custom_rules:
                            type: array
                            items:
                              type: string
                          custom_rule_sids:
                            type: array
                            items:
                              type: string
  /ConfigurationManager/updateConfiguration:
    put:
      tags:
        - ConfigurationManager
      summary: "Security Orchestrator updates the current configuration."
      description: "Update an existing deployment configuration. For snort3, this endpoint also supports rules_action add/remove/replace."
      operationId: "soUpdateConfiguration"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                config_id:
                  type: string
                  description: "Stable identifier of the deployment to update."
                configuration:
                  type: object
                  additionalProperties: true
                  description: "Optional partial env var overrides."
                rules_action:
                  type: string
                  enum: [add, remove, replace]
                  description: "snort3-only action for mutating custom rules."
                rules:
                  type: array
                  items:
                    type: string
                  description: "snort3-only list of custom rules. Required for add and replace."
                rule_sids:
                  type: array
                  items:
                    type: string
                  description: "snort3-only list of custom rule SIDs to remove. Required for remove."
                include_default_rules:
                  type: boolean
                  description: "snort3-only flag accepted during replace to set whether default rules are used together with custom rules."
              required:
                - config_id

      responses:
        "200":
          description: "Proceeding to update the configuration"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                    description: "Response code indicating success"
                  message:
                    type: string
                    example: "Proceeding to update the configuration"
                    description: "Response message confirming the success of the request."
...